제목: jwt 모듈 리펙토링 3 내용: jwt 모듈 리펙토링 3       - jwt 모듈화 추가:

+++ backend/com/jwt/middleware.py
@@ -1,8 +1,7 @@
 # -*- coding: utf-8 -*-
 from rest_framework import status
 
-from backend.com.jwt.handler import Handler
-from backend.com.jwt.salt import HttpSalt
+from backend.com.jwt.handler import Token, jwt_salt
 
 
 class JSONWebTokenMiddleware:
@@ -11,11 +10,20 @@ class JSONWebTokenMiddleware:
 
     def __call__(self, request):
         response = self.get_response(request)
+        return self.jwt_response(request, response)
+
+    @staticmethod
+    def jwt_response(request, response):
+        if getattr(request.user, 'is_staff', None):
+            return response
+
+        if getattr(request.user, 'is_superuser', None):
+            return response
 
         if getattr(request.user, 'is_authenticated', None):
-            handler = Handler(salt=HttpSalt(request))
-            payload = handler.get_jwt_payload(request.user.__dict__)
-            response.set_cookie('token', value=handler.jwt_encode_payload(payload), httponly=True, samesite='Lax')
+            token = Token(salt=jwt_salt(request))
+            payload = token.get_jwt_payload(request.user.__dict__)
+            response.set_cookie('token', value=token.jwt_encode_payload(payload), httponly=True, samesite='Lax')
 
         if response.status_code == status.HTTP_401_UNAUTHORIZED:
             response.delete_cookie('token')
